Add files via upload
The project aims at downloading reports from the SEC (Securities and Exchange Commission) website for all the companies. Python Edgar library is used to get the details of all the reports and SEC API is used to download the reports. To automate the downloading Selenium library is used. From the Edgar index page the report link is scraped the using BeautifulSoup library.